+++ src/components/Form/FormRangeSlider.tsx
@@ -1,6 +1,12 @@
 import classNames from 'classnames';
 import MultiRangeSlider, { ChangeResult } from 'multi-range-slider-react';
-import { ForwardedRef, forwardRef, ReactNode } from 'react';
+import {
+  ForwardedRef,
+  forwardRef,
+  ReactNode,
+  useEffect,
+  useState,
+} from 'react';
 import { FieldValues, Path, useFormContext } from 'react-hook-form';
 import { FormLabel } from './FormLabel';
 
@@ -9,6 +15,8 @@ export interface FormRangeSliderProps<Values> {
   barInnerColor?: string;
   children?: ReactNode;
   className?: string;
+  defaultMinValue?: number;
+  defaultMaxValue?: number;
   labelText?: string;
   min: number;
   max: number;
@@ -30,6 +38,8 @@ export const FormRangeSlider = forwardRef(
       barInnerColor,
       children,
       className,
+      defaultMinValue,
+      defaultMaxValue,
       labelText,
       min,
       max,
@@ -45,7 +55,14 @@ export const FormRangeSlider = forwardRef(
     }: FormRangeSliderProps<Values>,
     ref: ForwardedRef<HTMLDivElement>,
   ) => {
+    const [minValue, setMinValue] = useState(defaultMinValue ?? min);
+    const [maxValue, setMaxValue] = useState(defaultMaxValue ?? max);
     const { setValue } = useFormContext<Values>();
+    useEffect(() => {
+      setValue(name, [minValue, maxValue] as never, {
+        shouldDirty: true,
+      });
+    }, [minValue, maxValue]);
     return (
       <div className={classNames('form-control', className)}>
         <div className="flex justify-between">
@@ -65,13 +82,12 @@ export const FormRangeSlider = forwardRef(
           ruler={ruler}
           step={step}
           stepOnly={stepOnly}
-          minValue={min}
-          maxValue={max}
+          minValue={minValue}
+          maxValue={maxValue}
           onChange={onChange}
           onInput={event => {
-            setValue(name, [event.minValue, event.maxValue] as never, {
-              shouldDirty: true,
-            });
+            setMinValue(event.minValue);
+            setMaxValue(event.maxValue);
             onInput?.(event);
           }}
         />